requirements.txt
flask
pandas
scikit-learn
requests
python-whois
tldextract
gunicorn